@@ -1,25 +1,25 @@
-'use client'
-import Link from "next/link";
-import "../globals.css";
-import {Portal} from "@/app/components/Portal";
-import React, {useState} from "react";
-import SearchForm from "@/app/components/SearchForm";
-import {usePathname} from "next/navigation";
-import {Summary} from "@/app/components/Summuries";
+import "../../index.css";
+import {Portal} from "./Portal";
+import {useContext, useState} from "react";
+import SearchForm from "./SearchForm";
+import {Link, useLocation} from "react-router-dom";
+import NetworkInfoContext from "../../utils/context.tsx";
 
-export default function Header({networkOptions}: {networkOptions?: Summary[]}) {
+export default function Header() {
     const title = <p className="text-5xl text-white bg-[#27aab9] text-center p-7">BTP Message Explorer</p>;
     const [isOpen, setIsOpen] = useState(false);
+    const nMap = useContext(NetworkInfoContext);
     function handleOpenModal() {
         setIsOpen(true);
     }
     function handleCloseModal() {
         setIsOpen(false);
     }
+    if (Object.keys(nMap).length === 0) return <NaviBar handleOpenModal={handleOpenModal}/>;
     return (
         <>
             <div id={"modal-root"} className={"flex"}/>
-            {isOpen && <ModalSearchForm options={networkOptions} closeModal={handleCloseModal}/>}
+            {isOpen && <ModalSearchForm options={Object.keys(nMap)} closeModal={handleCloseModal}/>}
             <NaviBar handleOpenModal={handleOpenModal}/>
             {title}
         </>
@@ -30,7 +30,7 @@ function NaviBar({handleOpenModal}: {handleOpenModal: () => void}) {
     return (
         <nav className="border-gray-200 bg-[#27aab9]" aria-label="Main navigation">
             <div className="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4">
-                <Link href="/" className="items-center">
+                <Link to="/" className="items-center">
                     <span className="self-center text-3xl font-semibold whitespace-nowrap text-white">Home</span>
                 </Link>
                 <SearchButton openModal={handleOpenModal}/>
@@ -40,10 +40,10 @@ function NaviBar({handleOpenModal}: {handleOpenModal: () => void}) {
 }
 
 function SearchButton({openModal}: {openModal: () => void}) {
-    const path = usePathname();
+    const path = useLocation();
 
     const onSearchButton = () => {
-        if (path !== "/") {
+        if (path.pathname !== "/") {
             openModal();
         }
         const searchInput = document.querySelector("#searchInput") as HTMLInputElement
@@ -62,7 +62,7 @@ function SearchButton({openModal}: {openModal: () => void}) {
     )
 }
 
-function ModalSearchForm({options, closeModal}: {options?: Summary[], closeModal: () => void}) {
+function ModalSearchForm({options, closeModal}: {options: string[], closeModal: () => void}) {
     return (
         <Portal>
             <div onClick={() => {closeModal()}} className="justify-center fixed inset-0 z-10 bg-gray-100 bg-opacity-40">
