@@ -1,5 +1,4 @@
-'use client'
-import React, { ReactElement, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { createPortal } from "react-dom";
 
 export function Portal({ children }: { children: any }) {
