@@ -1,24 +1,23 @@
-'use client'
-import {MessageTable} from "@/app/messages/MessageTable";
-import Link from "next/link";
-import {BTPMessage} from "@/app/data/BTPMessage";
+import {MessageTable} from "./MessageTable";
+import {BTPMessage} from "../../data/BTPMessage";
 import {useEffect} from "react";
 import {QueryClient, QueryClientProvider, useQuery} from "react-query";
+import {Link} from "react-router-dom";
 
 const fetchData = async () => {
-    const response = await fetch(`${process.env.NEXT_PUBLIC_API_URI}/tracker/bmc?task=status&page=0&size=15&sort=created_at desc`);
+    const response = await fetch(`${import.meta.env.VITE_API_URL}/tracker/bmc?task=status&page=0&size=15&sort=created_at desc`);
     return await response.json();
 };
 export default function RefreshingTable() {
     const queryClient = new QueryClient();
     return (
         <QueryClientProvider client={queryClient}>
-            <Table/>
+            <TableCom/>
         </QueryClientProvider>
     )
 }
 
-function Table() {
+function TableCom() {
     const query = useQuery(
         ['data'],
         () => fetchData(),
@@ -36,7 +35,7 @@ function Table() {
     if (!query.data) return (
         <>
             <div className="w-full bg-gray-100 text-center text-[#27aab9] mt-7">
-                <Link href={"/messages"}>more ...</Link>
+                <Link to={"/messages"}>more ...</Link>
             </div>
         </>
     )
@@ -46,7 +45,7 @@ function Table() {
             <div className="w-full mt-7"></div>
             <MessageTable messages={messages}/>
             <div className="w-full bg-gray-100 text-center text-[#27aab9]">
-                <Link href={"/messages"}>more ...</Link>
+                <Link to={"/messages"}>more ...</Link>
             </div>
         </>
     );
